+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/config/ConfigService.java
@@ -31,6 +31,7 @@ public class ConfigService implements Service {
     private final String searchableAlarmTags;
     private final int metricsDataTTL;
     private final int recordDataTTL;
+    private final int persistentPeriod;
 
     public ConfigService(CoreModuleConfig moduleConfig) {
         this.gRPCHost = moduleConfig.getGRPCHost();
@@ -40,5 +41,6 @@ public class ConfigService implements Service {
         this.searchableAlarmTags = moduleConfig.getSearchableAlarmTags();
         this.metricsDataTTL = moduleConfig.getMetricsDataTTL();
         this.recordDataTTL = moduleConfig.getRecordDataTTL();
+        this.persistentPeriod = moduleConfig.getPersistentPeriod();
     }
 }