+++ docs/en/changes/changes.md
@@ -30,6 +30,10 @@
 * Fix miss set `LastUpdateTimestamp` that caused the metrics session to expire.
 * Rename MAL rule `spring-sleuth.yaml` to `spring-micrometer.yaml`.
 * Fix memory leak in Zipkin API.
+* Remove the dependency of `refresh_interval` of ElasticSearch indices from `elasticsearch/flushInterval` config. Now,
+  it uses `core/persistentPeriod` + 5s as `refresh_interval` for all indices instead.
+* Change `elasticsearch/flushInterval` to 5s(was 15s).
+* Optimize `flushInterval` of ElasticSearch BulkProcessor to avoid extra periodical flush in the continuous bulk streams. 
 
 #### OAP-Backend